auth_service/pyproject.toml
[project]
name = "auth-booking"
version = "0.1.0"
package-mode = false
description = ""
authors = [
    {name = "dusnila",email = "davidovich18062002@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi == 0.128.0",
    "uvicorn[standard] == 0.40.0",
    "sqlalchemy == 2.0.45",
    "alembic == 1.18.1",
    "asyncpg == 0.31.0",
    "pydantic-settings == 2.12.0",
    "python-dotenv == 1.2.1",
    "celery == 5.6.2",
    "redis == 4.6.0",
    "flower == 2.0.1",
    "passlib[bcrypt] == 1.7.4",
    "python-jose[cryptography] == 3.3.0",
    "email-validator == 2.3.0",
    "gunicorn == 23.0.0",
    "pendulum == 3.1.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
